+++ redcycle/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Account, Location, CommunityPost
+from .models.models import Account, Location, CommunityPost
 
 admin.site.register(Account)
 admin.site.register(Location)
